extract obstacle related members into PlateauBase abstract class
